Enhance application configuration and routing for tenant management
- Added error interceptor to handle HTTP errors globally and manage user sessions on 401 responses.
- Introduced tenant role guards to restrict access based on user roles for tenant administration.
- Updated routing to include new tenant management features, such as user and customer management pages.
- Enhanced application configuration to include the new error interceptor in HTTP client providers.
- Added models and services for customer and tenant user management, ensuring proper data handling and API interactions.

+++ server/POS.Application/Contracts/Products/ProductResponse.cs
@@ -18,7 +18,7 @@ public sealed class ProductResponse
 
     public decimal FinalPrice { get; init; }
 
-    public int Stock { get; init; }
+    public decimal Stock { get; init; }
 
     public decimal? LastCost { get; init; }
